+++ src/core/messaging/ACSClient.ts
@@ -407,8 +407,21 @@ class ACSClient {
     public async initialize(acsClientConfig: ACSClientConfig): Promise<void> {
         this.logger?.startScenario(ACSClientEvent.InitializeACSClient);
 
+        const tokenRefresher = async () => {
+            if (acsClientConfig.tokenRefresher) {
+                const token = await acsClientConfig.tokenRefresher();
+                return token;
+            }
+
+            return acsClientConfig.token;
+        };
+
         try {
-            this.tokenCredential = new AzureCommunicationTokenCredential(acsClientConfig.token);
+            this.tokenCredential = new AzureCommunicationTokenCredential({
+                token: acsClientConfig.token,
+                tokenRefresher, // tokenRefresher is executed when token found to be expired on performing HTTP calls
+                refreshProactively: true // Flag to whether refresh token 10 mins it expires
+            });
         } catch (error) {
             const exceptionDetails = {
                 response: 'CreateTokenCredentialFailure',